+++ hooks/helpers/useAndroidHealthInit.ts
